Add "Abstract Types" section to Readme
I had some difficulty understanding the following error message:
"Abstract type found, but no 'helpers' module specified for this engine"

I received this error because I had an Abstract Type in the object that I wanted
to serialize and deserialize but I didn't provide any implementations for Milk
to serialize or deserialize it.

I'm adding a section on "Abstract Types" to the Readme.md to document what I
learned from the issue and to hopefully help others and my future self.

See this thread for more information: https://github.com/jaredly/milk/issues/6